[Refactor] Consolidate TraitsUI/Qt helpers into microdrop_utils; add hover-scroll Enum editor.
Moves previously-scattered view helpers into centralized modules so they can be reused: `StatusIconEditor`, `SteppedSpinEditor`, and `RangeWithCustomViewHints` (renamed `RangeWithSteppedSpinViewHint`) migrate from `dropbot_status_and_controls/view_helpers.py` into `microdrop_utils/traitsui_qt_helpers.py`, and `RangeWithViewHints` / `SafeCancelTableHandler` migrate there from `microdrop_utils/pyface_helpers.py`. `_ScalingPixmapLabel` moves to `microdrop_utils/pyside_helpers.py`. All dependent import sites are updated.

Also adds a new `_MarqueeComboBox` widget and `FixedWidthEnumEditor` TraitsUI factory: non-editable combo boxes that ignore the widest-item sizeHint and marquee-scroll the collapsed text on mouse hover (holds at end until hover-out). Wired into the `dielectric_material` selector so long dielectric names no longer stretch the status panel; the display label is tightened to "Dielectric" and the placeholder entry reads "Pick a dielectric...".
